Add `Reset-GitHubOutput` (#424)
## Description

This pull request introduces a new `Reset-GitHubOutput` function to
reset the GitHub output file and adds corresponding tests to ensure its
functionality. The most important changes are as follows:

### New functionality:

*
[`src/functions/public/Commands/Reset-GitHubOutput.ps1`](diffhunk://#diff-6047f57121e0734ac0b09549e324b31d9ebcc1e73141125ea0f78d357b544fc1R1-R45):
Added the `Reset-GitHubOutput` function, which clears the contents of
the GitHub output file specified by `$env:GITHUB_OUTPUT` or a custom
path provided via the `-Path` parameter. Includes error handling for
missing or invalid paths.

### Testing updates:

*
[`tests/GitHub.Tests.ps1`](diffhunk://#diff-0b1d9ba345a583adce874126c13d6edd3f789416bb9c4db5df1e18af3608554cR313-R319):
Added a test case to verify that `Reset-GitHubOutput` clears the outputs
from the GitHub output file. This test sets an output using
`Set-GitHubOutput`, confirms its presence, and then validates that it is
cleared after calling `Reset-GitHubOutput`.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas