🪲 [Fix]: Fixes to app installations on enterprise organizations (#437)
## Description

This pull request introduces several fixes to improve the handling of
GitHub App installations and related functionality. The updates include
adding a new constructor for `GitHubAppInstallation`, refining object
iteration in private functions, and aligning property names in public
functions for consistency.

### Enhancements to `GitHubAppInstallation` class:

* Added a new constructor to the `GitHubAppInstallation` class to
support initialization with additional parameters, including `Target`
and `Type`. This allows for more detailed and flexible object creation.

### Updates to private functions for object iteration:

* Replaced `ForEach-Object` with explicit `foreach` loops in
`Get-GitHubAppInstallableOrganization`, improving readability and
maintainability of the code.
* Updated `Get-GitHubEnterpriseOrganizationAppInstallation` to use the
new `GitHubAppInstallation` constructor, enabling the inclusion of
`Organization` and `Type` parameters for installations.
* Refined `Get-GitHubOrganizationAppInstallation` to use `foreach` loops
for better clarity and aligned it with the new constructor for
`GitHubAppInstallation`.

### Consistency improvements in public functions:

* Updated `Connect-GitHubApp` to use the `Target` property instead of
`account.slug` for `InstallationName` and `Enterprise`, ensuring
consistent property naming. Additionally, fixed a casing issue for the
`Name` property.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas