🩹 [Patch]: Update `GitHubContext` formatting to show default context (#446)
## Description

This pull request introduces changes to improve the user experience and
functionality of GitHub-related PowerShell scripts. The updates include
enhancing the display format of context names and adding a new option
for listing organizations associated with the authenticated user.

- Fixes #445

### Display Format Improvements:

*
[`src/formats/GitHubContext.Format.ps1xml`](diffhunk://#diff-c7dd80cd4c4440ccbe24930842a2e172614dbfd79d31393d68852200eacc2c74L31-R37):
Updated the `TableColumnItem` for `Name` to use a `ScriptBlock` that
highlights the default context with a `">"` prefix while maintaining a
standard format for other contexts.

### Organization Listing Enhancements:

*
[`src/functions/public/Organization/Get-GitHubOrganization.ps1`](diffhunk://#diff-13a6b78f6ec32db51e7d358afb07f0fd1160f72201522a2d339f775b09d71274L116-R116):
Added a new case to the switch statement for listing organizations
associated with the authenticated user, improving clarity and
functionality.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas