🪲 [Fix]: Skip revoke token if token is expired (#488)
## Description

This pull request refactors the conditional logic in the
`Disconnect-GitHubAccount` function to improve readability and debugging
and that actually skips the revocation of the token if it is expired.

Refactoring and debugging improvements:

* Added a condition to check if the token is expired before running the
`Revoke-GitHubAppInstallationAccessToken`. The issue here was that the
function would fail if it was expired.
* Split the previous compound conditional into three distinct variables:
`$isNotGitHubToken`, `$isIATAuthType`, and `$isNotExpired` for clarity
and maintainability.
* Added `Write-Debug` statements for each condition to facilitate easier
troubleshooting and understanding of the script's flow.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas