🪲 [Fix]: Fixing the `Secrets` test for `PublicKey` for `codespaces` (#518)
## Description

This pull request updates the `Secrets.Tests.ps1` test suite to improve
how the `Get-GitHubPublicKey` command is tested for different
organization plans and makes a minor formatting change to a log message.

Test logic improvements for organization plan handling:

* Updated the `Get-GitHubPublicKey -Type codespaces` test to
conditionally expect a throw for organizations on the 'free' plan, and
to run the usual public key retrieval for other plans. This ensures the
test behaves correctly based on the organization's plan type.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas